Automatically rebuild affected files on compile flag changes.
This eliminates the need to carefully remove object files when we
run make with different compile flags.

While the core of this change is remarkably simple, it has
repercussions.  We have to split out flags for kern/init.c and net/*.c
because, now, changing DEFS will cause everything to be rebuilt (as it
should).  We also have to sleep after running make in the grade
scripts because, otherwise, we'll call make more than once per second,
which means it won't be able to detect mtime changes and will fail to
rebuild things (previously we forced it to rebuild exactly what we
knew we needed).