Awesomely better match_packet_seq for testinput and testoutput.
Use difflib to match up the real and expected packet sequences and
produce a failure message based on the full diff.  Previously, we
assumed they would align perfectly---which is almost certainly *not*
the case if something goes wrong---and reported on only the first
mismatched packet.

This also eliminates the dependency on tcpdump, instead reading the
pcap file directly.